***   WARNING  *** *** BIG CHANGE *** ***   WARNING  ***
- DOCX generation -
For docx the template is no more needed: a new document is created from scratch.
The name of styles used was changed, so the old template cannot be used.
The style are  made more consistent, all local (no more built-in)  and prefixed with 'OR-'
Any style defined in a template with the same name will be used instead of generated one. (* need testing)

The document (copy in memory) loaded from template will be emptyed before the use.
(The assumption is that first section is the body, but may be usefull check for header/footer)
The will be used the actual page width, minus left and right margins.
For a docx issue, those margins must be integer numbers: a workaround is
in place.

Added a page break after the title: make a cover for the report.

Refactored numpy code for chart generation: no more warnings.
Setted DPI and height of charts: better visualization (and higher definition).
Debug logging turned off

- CSV generation -
Issue a warning (and not an error) for using templates.
This makes consistent the interface.

***   WARNING  ***
*** BIG CHANGE ***
***   WARNING  ***